--- Deepgram.Tests/ClientTests/KeyClientTests.cs
@@ -1,10 +1,0 @@
-﻿namespace Deepgram.Tests.ClientTests;
-public class KeyClientTests
-{
-    [Fact]
-    public async void Should()
-    {
-
-
-    }
-}
