+++ Deepgram.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.0.31912.275
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Deepgram", "Deepgram\Deepgram.csproj", "{B8B59E33-2D73-4A27-8FC6-E464CB28357D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Deepgram.Tests", "Deepgram.Tests\Deepgram.Tests.csproj", "{12C80273-08DD-494C-B06D-DFC6D40B1D95}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Deepgram.Tests", "Deepgram.Test\Deepgram.Tests.csproj", "{A5934B7C-DE5C-4EEA-B567-43A18A81C1C1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{B8B59E33-2D73-4A27-8FC6-E464CB28357D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B8B59E33-2D73-4A27-8FC6-E464CB28357D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B8B59E33-2D73-4A27-8FC6-E464CB28357D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{12C80273-08DD-494C-B06D-DFC6D40B1D95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{12C80273-08DD-494C-B06D-DFC6D40B1D95}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{12C80273-08DD-494C-B06D-DFC6D40B1D95}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{12C80273-08DD-494C-B06D-DFC6D40B1D95}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A5934B7C-DE5C-4EEA-B567-43A18A81C1C1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A5934B7C-DE5C-4EEA-B567-43A18A81C1C1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A5934B7C-DE5C-4EEA-B567-43A18A81C1C1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A5934B7C-DE5C-4EEA-B567-43A18A81C1C1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
