+++ Deepgram/Deepgram.csproj
@@ -26,13 +26,9 @@
     <None Include="D:\Sources\deepgram\deepgram-dotnet-sdk\Deepgram\.editorconfig" />
   </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="7.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="7.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="7.0.0" />
+  <ItemGroup>  
     <PackageReference Include="Microsoft.Extensions.Logging" Version="7.0.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageReference Include="NSubstitute" Version="5.0.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />  
    <PackageReference Include="System.Threading.Channels" Version="7.0.0" />
   </ItemGroup>
 
@@ -44,4 +40,5 @@
       <_Parameter1>Deepgram.Tests</_Parameter1>
     </AssemblyAttribute>
   </ItemGroup>
+
 </Project>
